+++ packages/nextra/__test__/normalize-page.spec.ts
@@ -113,7 +113,7 @@ describe('normalize-page', () => {
     ])
 
     const result2 = normalizePages({
-      list: pageMap,
+      list: normalizePageMap(pageMap),
       route: '/1-level/2-level/foo'
     })
     expect(result2).toMatchSnapshot()
@@ -159,7 +159,7 @@ describe('normalize-page', () => {
     ])
 
     const { activeType, activeIndex, activeThemeContext } = normalizePages({
-      list: pageMap,
+      list: normalizePageMap(pageMap),
       route: '/1-level/not-exist'
     })
     expect({ activeType, activeIndex, activeThemeContext }).toEqual({
@@ -186,7 +186,7 @@ describe('normalize-page', () => {
     )
 
     const normalizedResult = normalizePages({
-      list: pageMap,
+      list: normalizePageMap(pageMap),
       route: '/one/two/qux'
     })
     expect(normalizedResult.docsDirectories).toMatchInlineSnapshot(`
@@ -280,7 +280,7 @@ describe('normalize-page', () => {
     const pageMap = await getPageMapForFixture('type-menu-should-contain-items')
 
     const normalizedResult = normalizePages({
-      list: pageMap,
+      list: normalizePageMap(pageMap),
       route: '/pagesOnly/one'
     })
     expect(
@@ -373,12 +373,12 @@ describe('normalize-page', () => {
     const pageMap = await getPageMapForFixture('pages-order-without-type-page')
 
     const normalizedResult = normalizePages({
-      list: pageMap,
+      list: normalizePageMap(pageMap),
       route: '/docs/bar'
     })
-    const { flatDirectories } = normalizedResult
-    expect(flatDirectories[0].name).toBe('_')
-    expect(flatDirectories[1].route).toBe('/docs/bar')
-    expect(flatDirectories[2].route).toBe('/foo')
+    const { docsDirectories } = normalizedResult
+    expect(docsDirectories[0].name).toBe('_')
+    expect(docsDirectories[1].route).toBe('/docs/bar')
+    expect(docsDirectories[3].route).toBe('/foo')
   })
 })