+++ packages/nextra/src/client/normalize-pages.ts
@@ -50,7 +50,6 @@ export type Item = (MdxFile | FolderWithoutChildren) & {
   type: string
   children: Item[]
   display?: Display
-  withIndexPage?: boolean
   theme?: PageTheme
   isUnderCurrentDocsTree?: boolean
 }
@@ -63,7 +62,6 @@ export type PageItem = (MdxFile | FolderWithoutChildren) & {
   children?: PageItem[]
   firstChildRoute?: string
   display?: Display
-  withIndexPage?: boolean
   isUnderCurrentDocsTree?: boolean
 }
 
@@ -77,7 +75,6 @@ type DocsItem = (MdxFile | FolderWithoutChildren) & {
   type: string
   children: DocsItem[]
   firstChildRoute?: string
-  withIndexPage?: boolean
   isUnderCurrentDocsTree?: boolean
 }
 
@@ -97,7 +94,6 @@ type NormalizedResult = {
   activeThemeContext: PageTheme
   activePath: Item[]
   directories: Item[]
-  flatDirectories: Item[]
   docsDirectories: DocsItem[]
   flatDocsDirectories: DocsItem[]
   topLevelNavbarItems: (PageItem | MenuItem)[]
@@ -118,9 +114,7 @@ export function normalizePages({
 }): NormalizedResult {
   // All directories
   // - directories: all directories in the tree structure
-  // - flatDirectories: all directories in the flat structure, used by search and footer navigation
   const directories: Item[] = []
-  const flatDirectories: Item[] = []
 
   // Docs directories
   const docsDirectories: DocsItem[] = []
@@ -129,16 +123,10 @@ export function normalizePages({
   // Page directories
   const topLevelNavbarItems: (PageItem | MenuItem)[] = []
 
-  const meta =
-    'data' in list[0]
-      ? (list[0].data as Record<string, Record<string, any> | undefined>)
-      : {}
+  const meta = 'data' in list[0] ? (list[0].data as MetaType) : {}
   // Normalize items based on files and _meta.json.
   const items = ('data' in list[0] ? list.slice(1) : list) as (
-    | (Folder & {
-        withIndexPage?: true
-        frontMatter: FrontMatter
-      })
+    | (Folder & { frontMatter?: FrontMatter })
     | MdxFile
   )[]
 
@@ -255,7 +243,7 @@ export function normalizePages({
               flatDocsDirectories.length + normalizedChildren.activeIndex
             break
         }
-        if ((currentItem as Item).withIndexPage && type === 'doc') {
+        if ('frontMatter' in currentItem && type === 'doc') {
           activeIndex++
         }
       }
@@ -268,11 +256,11 @@ export function normalizePages({
           docsDirectories.push(...normalizedChildren.docsDirectories)
 
           // If it's a page with children inside, we inject itself as a page too.
-          if (normalizedChildren.flatDirectories.length) {
-            const route = findFirstRoute(normalizedChildren.flatDirectories)
+          if (normalizedChildren.flatDocsDirectories.length) {
+            const route = findFirstRoute(normalizedChildren.flatDocsDirectories)
             if (route) pageItem.firstChildRoute = route
             topLevelNavbarItems.push(pageItem)
-          } else if (pageItem.withIndexPage) {
+          } else if ('frontMatter' in pageItem) {
             topLevelNavbarItems.push(pageItem)
           }
 
@@ -280,19 +268,17 @@ export function normalizePages({
         case 'doc':
           docsItem.children.push(...normalizedChildren.docsDirectories)
           // Itself is a doc page.
-          if (item.withIndexPage && display !== 'children') {
+          if ('frontMatter' in item && display !== 'children') {
             flatDocsDirectories.push(docsItem)
           }
       }
 
-      flatDirectories.push(...normalizedChildren.flatDirectories)
       flatDocsDirectories.push(...normalizedChildren.flatDocsDirectories)
       item.children.push(...normalizedChildren.directories)
     } else {
       if (isHidden) {
         continue
       }
-      flatDirectories.push(item)
       switch (type) {
         case 'page':
         case 'menu':
@@ -343,7 +329,6 @@ export function normalizePages({
     activeThemeContext,
     activePath,
     directories,
-    flatDirectories,
     docsDirectories,
     flatDocsDirectories,
     topLevelNavbarItems