+++ packages/nextra-theme-blog/CHANGELOG.md
@@ -169,6 +169,32 @@
 - Updated dependencies [fec399a]
   - nextra@3.1.0-canary.0
 
+## 3.1.0
+
+### Patch Changes
+
+- Updated dependencies [035fe48]
+- Updated dependencies [8e9767e]
+- Updated dependencies [fec399a]
+- Updated dependencies [c002118]
+  - nextra@3.1.0
+
+## 3.1.0-canary.1
+
+### Patch Changes
+
+- Updated dependencies [035fe48]
+- Updated dependencies [8e9767e]
+- Updated dependencies [c002118]
+  - nextra@3.1.0-canary.1
+
+## 3.1.0-canary.0
+
+### Patch Changes
+
+- Updated dependencies [fec399a]
+  - nextra@3.1.0-canary.0
+
 ## 3.0.15
 
 ### Patch Changes